# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JuneXDeployment API — Discord bot hosting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all bot templates
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new template (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
...
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a single template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update a template (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateUpdate"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a template (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Template:
      type: object
      required: [id, name, description, githubRepo, category, appJson, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        githubRepo:
          type: string
        thumbnail:
          type: ["string", "null"]
        category:
          type: string
        appJson:
          type: object
        createdAt:
          type: string
    TemplateInput:
      type: object
      required: [name, description, githubRepo, category, appJson]
      properties:
        name:
          type: string
        description:
          type: string
        githubRepo:
          type: string
        thumbnail:
          type: ["string", "null"]
        category:
          type: string
        appJson:
          type: object
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        githubRepo:
          type: string
        thumbnail:
          type: ["string", "null"]
        category:
          type: string
        appJson:
          type: object
...