# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateAppJson } from './templateAppJson';

export interface Template {
  id: number;
  name: string;
  description: string;
  githubRepo: string;
  /** @nullable */
  thumbnail?: string | null;
  category: string;
  appJson: TemplateAppJson;
  createdAt: string;
}