# lib/api-zod/src/generated/types/templateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateInputAppJson } from './templateInputAppJson';

export interface TemplateInput {
  name: string;
  description: string;
  githubRepo: string;
  /** @nullable */
  thumbnail?: string | null;
  category: string;
  appJson: TemplateInputAppJson;
}