# lib/api-zod/src/generated/types/templateUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * JuneXDeployment API — Discord bot hosting platform
 * OpenAPI spec version: 0.1.0
 */
import type { TemplateUpdateAppJson } from './templateUpdateAppJson';

export interface TemplateUpdate {
  name?: string;
  description?: string;
  githubRepo?: string;
  /** @nullable */
  thumbnail?: string | null;
  category?: string;
  appJson?: TemplateUpdateAppJson;
}